quote component is moved to the whatwedopage folder under components
